build: job worker test source generator
